exercise 1 done
NOTE:
i pushed module 3.1 with this exact description. please ignore that one

exercise 2, i understand how to do it
but when trying to make a page for each book using id, i can make the first book (id: 101) its own page, but breaks when input the second book's id (id: 102). i was stumped but i need to move on

exercise 3 was simple enough, and didnt bother to do it